+++ lib/deqm_test_kit/collect_data_v1.rb
@@ -1,97 +1,12 @@
 # frozen_string_literal: true
 
 require 'json'
+require_relative '../utils/collect_data_utils'
 
 module DEQMTestKit
   # tests for $collect-data (DEQM UV v1.0.0)
   # rubocop:disable Metrics/ClassLength
   class CollectDataV1 < Inferno::TestGroup
-    # module for shared code for $collect-data assertions and requests
-    module CollectDataHelpers
-      def selected_measure_url(custom_url:, url:, input_title: 'Measure URL',
-                               custom_input_title: 'Custom Measure URL')
-        return url unless url == 'Other'
-
-        custom_url = custom_url.to_s.strip
-
-        assert custom_url.length.positive?,
-               "#{custom_input_title} is required when \"#{input_title}\" is \"Other\"."
-
-        custom_url
-      end
-
-      def selected_additional_measure_url(custom_url:, url:)
-        selected_measure_url(
-          custom_url:,
-          url:,
-          input_title: 'Measure URL for additional Measure',
-          custom_input_title: 'Custom Additional Measure URL'
-        )
-      end
-
-      def selected_measure_urls
-        [
-          selected_measure_url(custom_url: custom_measure_url, url: measure_url),
-          selected_additional_measure_url(custom_url: custom_additional_measure_url, url: additional_measure_url)
-        ]
-      end
-
-      def validate_parameters_contains_bundles(parameters, measure_count)
-        assert parameters.parameter.is_a?(Array), 'Expected Parameters.parameter to be an array'
-        assert parameters.parameter.any?, 'Expected at least one parameter entry in Parameters resource'
-
-        parameters.parameter.each do |param|
-          assert param.resource.is_a?(FHIR::Bundle), 'Expected parameter.resource to be a Bundle'
-          validate_bundles_contain_measure_report(param.resource, measure_count)
-        end
-      end
-
-      def validate_number_of_bundles(parameters, bundle_count)
-        assert parameters.parameter.length == bundle_count,
-               "Expected #{bundle_count} Bundle(s), got #{parameters.parameter.length}"
-      end
-
-      def validate_bundles_contain_measure_report(bundle, measure_count)
-        assert bundle.entry.is_a?(Array), 'Expected Bundle.entry to be an array'
-        assert bundle.entry.any?, 'Expected at least one entry in the Bundle'
-
-        measure_reports = bundle.entry.map(&:resource).grep(FHIR::MeasureReport)
-        assert measure_reports.length == measure_count,
-               "Expected #{measure_count} MeasureReport(s), got #{measure_reports.length}"
-      end
-
-      def collect_data_body(period_start:, period_end:, measure_urls:, patient_id: nil) # rubocop:disable Metrics/MethodLength
-        parameters = measure_urls.map do |url|
-          {
-            name: 'measureUrl',
-            valueCanonical: url
-          }
-        end
-
-        if patient_id
-          parameters << {
-            name: 'subject',
-            valueString: "Patient/#{patient_id}"
-          }
-        end
-
-        parameters << {
-          name: 'periodStart',
-          valueDate: period_start
-        }
-
-        parameters << {
-          name: 'periodEnd',
-          valueDate: period_end
-        }
-
-        {
-          resourceType: 'Parameters',
-          parameter: parameters
-        }
-      end
-    end
-
     id :collect_data_v1
     title '$collect-data'
     description 'Ensure FHIR server can perform the $collect-data operation'
@@ -131,7 +46,7 @@ module DEQMTestKit
     }
 
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'GET Measure/$collect-data with one measureUrl, periodStart, periodEnd'
       id 'collect-data-one-measure-get'
@@ -161,7 +76,7 @@ module DEQMTestKit
     end
 
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'POST Measure/$collect-data with one measureUrl, periodStart, periodEnd'
       id 'collect-data-one-measure-post'
@@ -190,7 +105,7 @@ module DEQMTestKit
     end
 
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'GET Measure/$collect-data with two measureUrls, periodStart, periodEnd'
       id 'collect-data-two-measure-get'
@@ -223,7 +138,7 @@ module DEQMTestKit
     end
 
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'POST Measure/$collect-data with two measureUrls, periodStart, periodEnd'
       id 'collect-data-two-measure-post'
@@ -255,7 +170,7 @@ module DEQMTestKit
     end
 
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'GET Measure/$collect-data with one measureUrl, periodStart, periodEnd, and subject=Patient/patientId'
       id 'collect-data-one-measure-get-subject-patient'
@@ -290,7 +205,7 @@ module DEQMTestKit
     end
 
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'POST Measure/$collect-data with one measureUrl, periodStart, periodEnd, and subject=Patient/patientId'
       id 'collect-data-one-measure-post-subject-patient'
@@ -325,7 +240,7 @@ module DEQMTestKit
     end
 
     test do # rubocop:disable Metrics/BlockLength
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'GET Measure/$collect-data with two measureUrls, periodStart, periodEnd, and subject=Patient/patientId'
       id 'collect-data-two-measure-get-subject-patient'
@@ -362,7 +277,7 @@ module DEQMTestKit
     end
 
     test do # rubocop:disable Metrics/BlockLength
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'POST Measure/$collect-data with two measureUrls, periodStart, periodEnd, and subject=Patient/patientId'
       id 'collect-data-two-measure-post-subject-patient'
@@ -400,7 +315,7 @@ module DEQMTestKit
     # GET Measure/$collect-data with measureUrl and periodStart in request parameters returns 400 and
     # FHIR Operation Outcome when periodEnd was missing.
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'GET Measure/$collect-data missing periodEnd returns HTTP 400 and OperationOutcome'
       id 'collect-data-missing-period-end-get-fail'
@@ -431,7 +346,7 @@ module DEQMTestKit
     # POST Measure/$collect-data with measureUrl and periodStart in request body returns 400 and
     # FHIR Operation Outcome when periodEnd was missing.
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'POST Measure/$collect-data missing periodEnd returns HTTP 400 and OperationOutcome'
       id 'collect-data-missing-period-end-post-fail'
@@ -462,7 +377,7 @@ module DEQMTestKit
     # GET Measure/$collect-data with measureUrl and periodEnd in request parameters returns 400 and
     # FHIR Operation Outcome when periodStart was missing.
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'GET Measure/$collect-data missing periodStart returns HTTP 400 and OperationOutcome'
       id 'collect-data-missing-period-start-get-fail'
@@ -493,7 +408,7 @@ module DEQMTestKit
     # POST Measure/$collect-data with measureUrl and periodEnd in request body returns 400 and
     # FHIR Operation Outcome when periodStart was missing.
     test do
-      include CollectDataHelpers
+      include CollectDataUtils
 
       title 'POST Measure/$collect-data missing periodStart returns HTTP 400 and OperationOutcome'
       id 'collect-data-missing-period-start-post-fail'
