+++ spec/deqm_test_kit/v1.0.0/collect_data_spec.rb
@@ -1,55 +1,19 @@
 # frozen_string_literal: true
 
+require_relative '../../utils/collect_data_spec_utils'
+
 INVALID_MEASURE_ID = 'INVALID_MEASURE_ID'
 INVALID_START_DATE = 'INVALID_START_DATE'
 
 RSpec.describe DEQMTestKit::CollectDataV1 do
+  include DEQMTestKit::CollectDataSpecUtils
+
   let(:suite) { Inferno::Repositories::TestSuites.new.find('deqm_v100') }
   let(:group) { suite.groups[2] }
   let(:session_data_repo) { Inferno::Repositories::SessionData.new }
   let(:test_session) { repo_create(:test_session, test_suite_id: suite.id) }
   let(:url) { 'http://example.com/fhir' }
   let(:error_outcome) { FHIR::OperationOutcome.new(issue: [{ severity: 'error' }]) }
-
-  # Helper method to create a valid Parameters resource
-  def create_parameters_response(measure_urls:) # rubocop:disable Metrics/MethodLength
-    measure_reports = measure_urls.map do |url|
-      FHIR::MeasureReport.new(
-        status: 'complete', type: 'data-collection',
-        measure: url,
-        period: { start: '2019-01-01', end: '2019-12-31' }
-      )
-    end
-
-    bundle = FHIR::Bundle.new(type: 'transaction', entry: measure_reports.map { |mr| { resource: mr } })
-
-    FHIR::Parameters.new(parameter:
-      {
-        name: 'return',
-        resource: bundle
-      })
-  end
-
-  def create_parameters_request(measure_urls:, period_start:, period_end:, patient_id: nil)
-    parameters = measure_urls.map { |measure_url| { name: 'measureUrl', valueCanonical: measure_url } }
-    parameters << { name: 'subject', valueString: "Patient/#{patient_id}" } if patient_id
-    parameters << { name: 'periodStart', valueDate: period_start }
-    parameters << { name: 'periodEnd', valueDate: period_end }
-
-    {
-      resourceType: 'Parameters',
-      parameter: parameters
-    }
-  end
-
-  def run(runnable, inputs = {})
-    test_run_params = { test_session_id: test_session.id }.merge(runnable.reference_hash)
-    test_run = Inferno::Repositories::TestRuns.new.create(test_run_params)
-    inputs.each do |name, value|
-      session_data_repo.save(test_session_id: test_session.id, name:, value:, type: 'text')
-    end
-    Inferno::TestRunner.new(test_session:, test_run:).run(runnable)
-  end
 
   describe 'GET Measure/$collect-data with one measureUrl and required params' do
     let(:test) { test_by_id(group, 'collect-data-one-measure-get') }
